+++ src/main/java/racingcar/domain/CarName.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain;
 
 import racingcar.exception.CarNameInputException;
