+++ src/main/java/racingcar/domain/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain;
 
 import racingcar.util.MathUtil;
